refactor(test): P5-1 follow-up — extract HTML ARIA + querySelector cluster
Continue test_widget_bridge.cpp split. Extracts the 509-line Wave 3
html bundle (pulp #1476: html.2 + html.3) into a sibling TU:

  1. ARIA attribute routing — aria-label / role / aria-pressed /
     -checked / -disabled / -hidden flow through the html-compat shim
     into View accessibility slots that the macOS NSAccessibility
     bridge consumes; setAttribute / removeAttribute round-trip;
     mount-deferred attributes replay on appendChild.

  2. querySelector / querySelectorAll — selector engine accepts tag /
     .class / #id / attribute selectors, colons inside attribute
     brackets, descendant ( ) and child (>) combinators, :hover /
     :disabled / :checked / :enabled / :not() pseudo-classes, and
     structural :first-child / :last-child / :nth-child / :only-child
     / :empty.

* test_widget_bridge.cpp: 10,700 → 10,191 (-509)
* New target pulp-test-widget-bridge-html-aria: 551 lines, 13
  TEST_CASEs, 66 assertions.

This is commit 2 of 2 in a bundled-cuts PR (SVG cluster shipped in the
preceding commit; single version bump comes next).
