+++ test/CMakeLists.txt
@@ -1787,6 +1787,13 @@ pulp_add_test_suite(pulp-test-widget-bridge-canvas2d-wave2 LIBRARIES pulp::view)
 # parser regression (Spectr PEAK / AVG / BOTH / OFF analyzer icons).
 pulp_add_test_suite(pulp-test-widget-bridge-svg LIBRARIES pulp::view)
 
+# Widget bridge — HTML ARIA + querySelector (P5-1 follow-up, Wave 3
+# html bundle). aria-label / role flow into View accessibility slots;
+# document.querySelector accepts attribute selectors + combinators +
+# :hover / :disabled / :checked / :enabled / :not / :first-child /
+# :nth-child / :empty pseudo-classes.
+pulp_add_test_suite(pulp-test-widget-bridge-html-aria LIBRARIES pulp::view)
+
 # Autonomous Spectr regression suite (pulp-internal task #63, follow-up to
 # pulp #1792). Composes the six [contract] invariants from
 # test_widget_bridge.cpp into a Spectr-shaped mini-scenario so a future
